pyproject.toml
[project]
name = "kord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "fastapi[standard]>=0.135.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "lxml>=5.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sarvamai>=0.1.25",
    "supabase>=2.0.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.26.0",
]
